refactor rfid -> smart card handler for mqtt when record smart card plan is refactor exception code add constraint for smart card, smart card only can be have with santri or employee not both

+++ internal/handler/parent_handler.go
@@ -128,7 +128,7 @@ func (h *parentHandler) ListParentHandler(c *gin.Context) {
 
 	pagination := model.Pagination{
 		CurrentPage:  listParentRequest.Page,
-		TotalPages:   (count + listParentRequest.Limit - 1) / listParentRequest.Limit,
+		TotalPages:   int32((count + int64(listParentRequest.Limit) - 1) / int64(listParentRequest.Limit)),
 		TotalItems:   count,
 		ItemsPerPage: listParentRequest.Limit,
 	}
